feat(readme): introduce README and rename elements
Rename ConnectionsInfo into SessionInfo, since it represents the session
information between the device and the Edgehog bridge.
Also rename the `connections.rs` module into `collection.rs`,
since it reprtesents a collection of Connection elements and defines the
methods to handle it.
