+++ edgehog-device-runtime-forwarder/src/collection.rs
